@@ -9,34 +9,36 @@
     "build": "ng build --prod --configuration=production",
     "pretest": "",
     "test": "echo 'Test suite should be run with \"`npm run tests` or npm `run tests-ci`\"'",
-    "tests": "ng test",
+    "tests": "ng test --watch=false",
     "tests-ci": "ng test --watch=false",
     "lint": "ng lint --format=stylish",
-    "e2e": "ng e2e",
-    "postinstall": "ng build --prod --configuration=production"
+    "e2e": "ng e2e"
   },
+  "pre-push": [
+    "lint"
+  ],
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
-    "@angular/http": "^6.1.10",
-    "@angular/material": "6.4.7",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/platform-server": "^6.1.10",
-    "@angular/router": "^6.1.10",
-    "@ng-bootstrap/ng-bootstrap": "^3.3.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^5.3.0",
+    "@ng-select/ng-select": "^3.7.3",
     "@types/geojson": "^7946.0.2",
     "@types/leaflet": "^1.2.6",
     "@types/leaflet.markercluster": "^1.0.3",
-    "bootstrap": "^4.3.1",
+    "bootstrap": "^4.4.1",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.5.3",
+    "core-js": "^3.1.3",
     "hammerjs": "^2.0.8",
     "intl": "^1.2.5",
     "jquery": "^3.4.1",
@@ -46,41 +48,45 @@
     "moment": "^2.21.0",
     "moment-timezone": "^0.5.25",
     "mygovbc-bootstrap-theme": "^0.4.1",
+    "ng-zorro-antd": "8.5.2",
     "ng2-bootstrap-modal": "^1.0.1",
-    "ngx-chips": "2.0.2",
+    "ngx-chips": "2.1.0",
     "ngx-cookie-service": "1.0.10",
-    "ngx-page-scroll": "^5.0.1",
-    "ngx-pagination": "^3.3.1",
+    "ngx-page-scroll": "^7.0.1",
+    "ngx-page-scroll-core": "^7.0.1",
     "ngx-text-overflow-clamp": "0.0.1",
     "popper.js": "^1.14.1",
-    "rxjs": "^6.5.2",
-    "rxjs-compat": "^6.5.2",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.29"
+    "rxjs": "^6.5.4",
+    "rxjs-compat": "^6.5.4",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
     "8": "0.0.1",
-    "@angular-devkit/build-angular": "^0.8.9",
-    "@angular-devkit/core": "0.0.22",
-    "@angular-devkit/schematics": "0.0.40",
-    "@angular/cli": "^6.2.9",
+    "@angular-devkit/build-angular": "^0.803.23",
+    "@angular-devkit/core": "^8.3.25",
+    "@angular-devkit/schematics": "^8.3.25",
+    "@angular/cli": "^8.3.25",
     "@types/arcgis-js-api": "4.6.0",
     "@types/jasmine": "2.8.6",
     "@types/lodash": "4.14.106",
-    "@types/node": "^8.10.51",
+    "@types/node": "^10.17.17",
     "codelyzer": "^4.5.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "node": "^8.16.0",
-    "protractor": "^6.0.0",
-    "ts-node": "2.0.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "~3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
+    "node": "^10.19.0",
+    "pre-push": "^0.1.1",
+    "protractor": "~5.4.3",
+    "puppeteer": "^2.1.1",
+    "ts-node": "^7.0.1",
     "tslint": "^5.19.0",
-    "typescript": "^2.9.2"
+    "typescript": "3.4.5",
+    "tree-kill": "^1.2.2"
   }
 }
