@@ -1,18 +1,27 @@
-import { Component, Input, OnInit, ComponentFactoryResolver, OnDestroy, ViewChild, Output, EventEmitter, SimpleChanges, OnChanges } from '@angular/core';
+import { Component, Input, OnInit, ComponentFactoryResolver, OnDestroy, ViewChild, Output, EventEmitter, SimpleChanges, OnChanges, ViewEncapsulation } from '@angular/core';
 
 import { TableDirective } from './table.directive';
 import { TableObject } from './table-object';
 import { TableComponent } from './table.component';
+import { Constants } from 'app/shared/utils/constants';
 
 @Component({
   selector: 'app-table-template',
   templateUrl: './table-template.component.html',
-  styleUrls: ['./table-template.component.scss']
+  styleUrls: ['./table-template.component.scss'],
+  encapsulation: ViewEncapsulation.None,
 })
 export class TableTemplateComponent implements OnInit, OnChanges, OnDestroy {
   @Input() data: TableObject;
   @Input() columns: any[];
-  @ViewChild(TableDirective) tableHost: TableDirective;
+  @Input() pageSizeArray: number[];
+  @Input() activePageSize: number;
+  @Input() activePage: number = Constants.tableDefaults.DEFAULT_CURRENT_PAGE;
+  @Input() hidePager = false;
+  @Input() showMoreLoader = false;
+  @Input() showMoreIncrement: number = Constants.tableDefaults.DEFAULT_SHOW_MORE_INCREMENT;
+  @Input() showCountAtTop = true;
+  @ViewChild(TableDirective, {static: true}) tableHost: TableDirective;
 
   @Output() onPageNumUpdate: EventEmitter<any> = new EventEmitter();
   @Output() onSelectedRow: EventEmitter<any> = new EventEmitter();
@@ -25,6 +34,13 @@ export class TableTemplateComponent implements OnInit, OnChanges, OnDestroy {
 
   ngOnInit() {
     this.loadComponent();
+    this.activePageSize = parseInt(this.data.paginationData.pageSize, 10);
+    const pageSizeTemp = [10, 25, 50, 100, parseInt(this.data.paginationData.totalListItems, 10)];
+    this.pageSizeArray = pageSizeTemp.filter(function(el: number) { return el >= 10; });
+    this.pageSizeArray.sort(function(a: number, b: number) { return a - b });
+    if (this.activePage !== parseInt(this.data.paginationData.currentPage, 10)) {
+      this.activePage = parseInt(this.data.paginationData.currentPage, 10);
+    }
   }
 
   ngOnChanges(changes: SimpleChanges) {
@@ -33,6 +49,7 @@ export class TableTemplateComponent implements OnInit, OnChanges, OnDestroy {
       this.data.component = changes['data'].currentValue.component;
       this.data.data = changes['data'].currentValue.data;
       this.data.paginationData = changes['data'].currentValue.paginationData;
+      this.data.extraData = changes['data'].currentValue.extraData;
       this.column = changes['data'].currentValue.paginationData.sortBy;
       this.loadComponent();
     }
@@ -70,5 +87,9 @@ export class TableTemplateComponent implements OnInit, OnChanges, OnDestroy {
 
   updatePageNumber(pageNum) {
     this.onPageNumUpdate.emit(pageNum);
+  }
+  updatePageSize(pageSize) {
+    this.data.paginationData.pageSize = pageSize;
+    this.onPageNumUpdate.emit(1);
   }
 }
