@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { ContactComponent } from 'app/contact/contact.component';
+import { CACUnsubscribeComponent } from 'app/cac-unsubscribe/cac-unsubscribe.component';
 import { ProjectsComponent } from 'app/projects/projects.component';
 import { HomeComponent } from 'app/home/home.component';
 import { ProjectListComponent } from './projects/project-list/project-list.component';
@@ -11,13 +12,24 @@ import { ProcessComponent } from './process/process.component';
 import { ComplianceOversightComponent } from './compliance-oversight/compliance-oversight.component';
 import { SearchHelpComponent } from './search-help/search-help.component';
 import { NewsResolver } from './news/news-resolver.service';
-import { NotificationProjectsListComponent } from './notification-projects/notification-projects.component';
-import { NotificationProjectsResolver } from './notification-projects/notification-projects-resolver.service';
+import { ProjectNotificationsListComponent } from './project-notifications/project-notifications.component';
+import { ProjectNotificationsResolver } from './project-notifications/project-notifications-resolver.service';
+import { ProjectNotificationResolver } from './project-notifications/project-notification-resolver.service';
+import { CommentsComponent } from 'app/comments/comments.component';
+import { CommentsResolver } from 'app/comments/comments-resolver.service';
+import { SearchComponent } from 'app/search/search.component';
+import { SearchResolver } from './search/search-resolver.service';
+import { DocumentTableResolver } from './project/documents/project-document-table-rows/project-document-table-rows-resolver.service';
+
 
 const routes: Routes = [
   {
     path: 'contact',
     component: ContactComponent
+  },
+  {
+    path: 'cac-unsubscribe',
+    component: CACUnsubscribeComponent
   },
   {
     path: 'projects',
@@ -28,10 +40,23 @@ const routes: Routes = [
     component: ProjectListComponent
   },
   {
-    path: 'notification-projects',
-    component: NotificationProjectsListComponent,
+    path: 'project-notifications',
+    component: ProjectNotificationsListComponent,
     resolve: {
-      notificationProjects: NotificationProjectsResolver
+      projectNotifications: ProjectNotificationsResolver
+    }
+  },
+  {
+    path: 'pn/:projId/cp/:commentPeriodId',
+    redirectTo: 'pn/:projId/cp/:commentPeriodId/details',
+    pathMatch: 'full'
+  },
+  {
+    path: 'pn/:projId/cp/:commentPeriodId/details',
+    component: CommentsComponent,
+    resolve: {
+      commentPeriod: CommentsResolver,
+      project: ProjectNotificationResolver
     }
   },
   {
@@ -54,6 +79,14 @@ const routes: Routes = [
     component: ProcessComponent
   },
   {
+    path: 'search',
+    component: SearchComponent,
+    resolve: {
+      documents: SearchResolver,
+      documentsTableRows: DocumentTableResolver
+    }
+  },
+  {
     path: 'search-help',
     component: SearchHelpComponent
   },
@@ -73,7 +106,7 @@ const routes: Routes = [
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule],
-  providers: [NewsResolver, NotificationProjectsResolver]
+  providers: [NewsResolver, ProjectNotificationsResolver, ProjectNotificationResolver]
 })
 
 export class AppRoutingModule { }
