@@ -1,5 +1,4 @@
 import { Project } from './project';
-import { CommentPeriod } from './commentperiod';
 
 export class News {
   _id: number;
@@ -13,6 +12,7 @@ export class News {
   dateUpdated: string;
   contentUrl: string;
   documentUrl: string;
+  notificationName: string;
   constructor(obj?: any) {
     this._id = obj && obj._id || null;
     this.headline = obj && obj.headline || null;
@@ -24,6 +24,7 @@ export class News {
     this.dateAdded = obj && obj.dateAdded || null;
     this.dateUpdated = obj && obj.dateUpdated || null;
     this.contentUrl = obj && obj.contentUrl || null;
+    this.notificationName = obj && obj.notificationName || null;
     this.documentUrl = obj && obj.documentUrl || null;
   }
 }
