@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
 import 'rxjs/add/operator/switchMap';
@@ -7,25 +7,30 @@ import 'rxjs/add/operator/switchMap';
 import * as _ from 'lodash';
 
 import { SearchService } from 'app/services/search.service';
+import { StorageService } from 'app/services/storage.service';
 
 @Injectable()
 export class DocumentsResolver implements Resolve<Observable<object>> {
   private milestones: any[] = [];
   private authors: any[] = [];
   private types: any[] = [];
+  private projectPhases: any[] = []
 
   private filterForAPI: object = {};
 
   constructor(
-    private searchService: SearchService
+    private searchService: SearchService,
+    private storageService: StorageService
   ) { }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<object> {
+  resolve(route: ActivatedRouteSnapshot): Observable<object> {
     const projectId = route.parent.paramMap.get('projId');
-    const currentPage = route.params.currentPage ? route.params.currentPage : 1;
-    const pageSize = route.params.pageSize ? route.params.pageSize : 10;
-    const sortBy = route.params.sortBy && route.params.sortBy !== 'null' ? route.params.sortBy : '-datePosted';
-    const keywords = route.params.keywords;
+    let currentPage = route.params.currentPage ? route.params.currentPage : 1;
+    let pageSize = route.params.pageSize ? route.params.pageSize : 10;
+    let sortBy = route.params.sortBy ? route.params.sortBy : '-datePosted,+displayName';
+    const datePostedStart = route.params.hasOwnProperty('datePostedStart') &&  route.params.datePostedStart ? route.params.datePostedStart : null;
+    const datePostedEnd = route.params.hasOwnProperty('datePostedEnd') && route.params.datePostedEnd ? route.params.datePostedEnd : null;
+    const keywords = route.params.hasOwnProperty('keywords') ? route.params.keywords : '';
 
     // Get the lists first
     return this.searchService.getFullList('List')
@@ -42,6 +47,9 @@ export class DocumentsResolver implements Resolve<Observable<object>> {
               case 'doctype':
                 this.types.push({ ...item });
                 break;
+              case 'projectPhase':
+                this.projectPhases.push({ ...item });
+                break;
               default:
                 break;
             }
@@ -51,6 +59,28 @@ export class DocumentsResolver implements Resolve<Observable<object>> {
         // Validate the filter parameters being sent to the API
         this.setFilterFromParams(route.params);
 
+        let queryModifiers = { documentSource: 'PROJECT' };
+
+        if (datePostedStart !== null && datePostedEnd !== null) {
+          queryModifiers['datePostedStart'] = datePostedStart;
+          queryModifiers['datePostedEnd'] = datePostedEnd;
+        }
+
+        if (this.storageService && this.storageService.state[projectId]) {
+          let filterForUI = this.storageService.state[projectId].filterForUI
+          let tableParams = this.storageService.state[projectId].tableParams;
+
+          if (tableParams) {
+            currentPage = tableParams.currentPage;
+            pageSize = tableParams.pageSize;
+            sortBy = tableParams.sortBy ? tableParams.sortBy : sortBy;
+          }
+
+          if (filterForUI) {
+            this.setParamsFromFilters(filterForUI);
+          }
+        }
+
         return this.searchService.getSearchResults(
           keywords,
           'Document',
@@ -58,7 +88,7 @@ export class DocumentsResolver implements Resolve<Observable<object>> {
           currentPage,
           pageSize,
           sortBy,
-          { documentSource: 'PROJECT' },
+          queryModifiers,
           true,
           null,
           this.filterForAPI,
@@ -97,8 +127,20 @@ export class DocumentsResolver implements Resolve<Observable<object>> {
     this.paramsToCollectionFilter(params, 'milestone', this.milestones, '_id');
     this.paramsToCollectionFilter(params, 'documentAuthorType', this.authors, '_id');
     this.paramsToCollectionFilter(params, 'type', this.types, '_id');
+    this.paramsToCollectionFilter(params, 'projectPhase', this.projectPhases, '_id');
+  }
 
-    this.paramsToDateFilter(params, 'datePostedStart');
-    this.paramsToDateFilter(params, 'datePostedEnd');
+  setParamsFromFilters(filterForUI) {
+    this.collectionFilterToParams(filterForUI, 'milestone', '_id');
+    this.collectionFilterToParams(filterForUI, 'documentAuthorType', '_id');
+    this.collectionFilterToParams(filterForUI, 'type', '_id');
+    this.collectionFilterToParams(filterForUI, 'projectPhase', '_id');
+  }
+
+  collectionFilterToParams(filterForUI, name, identifyBy) {
+    if (filterForUI[name].length) {
+      const values = filterForUI[name].map(record => { return record[identifyBy]; });
+      this.filterForAPI[name] = values.join(',');
+    }
   }
 }
