@@ -3,7 +3,6 @@ import { Subject } from 'rxjs';
 import { TableComponent } from 'app/shared/components/table-template/table.component';
 import { TableObject } from 'app/shared/components/table-template/table-object';
 import { Router, ActivatedRoute } from '@angular/router';
-import { ApiService } from 'app/services/api';
 import { Utils } from 'app/shared/utils/utils';
 
 @Component({
@@ -19,18 +18,28 @@ export class DocumentTableRowsComponent implements OnInit, OnDestroy, TableCompo
 
   public documents: any;
   public paginationData: any;
+  public showFeatured = true;
   private lists: any[] = [];
+
+  public currentUrl: String = '';
+
   constructor(
     private _changeDetectionRef: ChangeDetectorRef,
     private route: ActivatedRoute,
-    private api: ApiService,
     private router: Router,
     private utils: Utils
-  ) { }
+  ) {
+    let currRoute = this.router.url.split(';')[0];
+    this.currentUrl = currRoute.substring(currRoute.lastIndexOf('/') + 1);
+   }
 
   ngOnInit() {
     this.documents = this.data.data;
     this.paginationData = this.data.paginationData;
+    if (this.data.extraData) {
+      this.showFeatured = this.data.extraData.showFeatured;
+    }
+
     this.route.data
       .takeUntil(this.ngUnsubscribe)
       .subscribe((res: any) => {
@@ -76,11 +85,11 @@ export class DocumentTableRowsComponent implements OnInit, OnDestroy, TableCompo
     let filename = item.documentFileName;
     let safeName = filename;
     try {
-      safeName = this.utils.encodeFilename(filename, true)
+      safeName = this.utils.encodeString(filename, true)
     } catch (e) {
       console.log('error:', e);
     }
-    window.open('/api/document/' + item._id + '/fetch/' + safeName, '_blank');
+    window.open('/api/public/document/' + item._id + '/download/' + safeName, '_blank');
   }
   ngOnDestroy() {
     this.ngUnsubscribe.next();
