@@ -1,8 +1,10 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
 import { ProjectService } from 'app/services/project.service';
+import { Org } from 'app/models/organization';
+import { DataQueryResponse } from 'app/models/api-response';
 
 @Injectable()
 export class PinsResolverService implements Resolve<Observable<object>> {
@@ -10,7 +12,7 @@ export class PinsResolverService implements Resolve<Observable<object>> {
     private projectService: ProjectService
   ) { }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<object> {
+  resolve(route: ActivatedRouteSnapshot): Observable<DataQueryResponse<Org>[]> {
     const projectId = route.parent.paramMap.get('projId');
     const pageNum = route.params.currentPage ? route.params.currentPage : 1;
     const pageSize = route.params.pageSize ? route.params.pageSize : 10;
