@@ -2,9 +2,6 @@ import { Component, Input, Output, OnInit, EventEmitter } from '@angular/core';
 
 import { TableComponent } from 'app/shared/components/table-template/table.component';
 import { TableObject } from 'app/shared/components/table-template/table-object';
-import { Router } from '@angular/router';
-import { ApiService } from 'app/services/api';
-import { Utils } from 'app/shared/utils/utils';
 
 @Component({
   selector: 'tbody[app-pins-table-rows]',
@@ -20,9 +17,6 @@ export class PinsTableRowsComponent implements OnInit, TableComponent {
   public paginationData: any;
 
   constructor(
-    private api: ApiService,
-    private router: Router,
-    private utils: Utils
   ) { }
 
   ngOnInit() {
@@ -42,14 +36,6 @@ export class PinsTableRowsComponent implements OnInit, TableComponent {
     this.selectedCount.emit(count);
   }
 
-  goToItem(item) {
-    // let filename = item.documentFileName;
-    // let safeName = filename;
-    // try {
-      // safeName = this.utils.encodeFilename(filename, true);
-    // } catch (e) {
-    //   console.log('error:', e);
-    // }
-    // window.open('/api/document/' + item._id + '/fetch/' + safeName, '_blank');
+  goToItem() {
   }
 }
