@@ -1,194 +1,173 @@
-import { Component, OnInit, AfterViewInit, OnDestroy, ElementRef } from '@angular/core';
-import * as L from 'leaflet';
-
+import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import { StorageService } from 'app/services/storage.service';
 import { Subject } from 'rxjs';
 import { ConfigService } from 'app/services/config.service';
-import { NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
+import { TableObject } from 'app/shared/components/table-template/table-object';
+import { TableParamsObject } from 'app/shared/components/table-template/table-params-object';
+import { TableTemplateUtils } from 'app/shared/utils/table-template-utils';
+import { SearchService } from 'app/services/search.service';
+import { DocumentTableRowsComponent } from '../documents/project-document-table-rows/project-document-table-rows.component';
+import { Document } from 'app/models/document';
 
 @Component({
   selector: 'app-project-details-tab',
   templateUrl: './project-details-tab.component.html',
-  styleUrls: ['./project-details-tab.component.scss'],
+  styleUrls: ['./project-details-tab-sm.component.scss', './project-details-tab-md-lg.component.scss'],
 })
-export class ProjectDetailsTabComponent implements OnInit, AfterViewInit, OnDestroy {
+export class ProjectDetailsTabComponent implements OnInit {
   public project;
   public commentPeriod = null;
-  public map: L.Map = null;
-  public appFG = L.featureGroup(); // group of layers for subject app
+  public selectedCount = 0;
+  public loading = true;
+  public featuredDocuments: Document[] = null;
+  public documentTableData: TableObject;
+  public tableParams: TableParamsObject = new TableParamsObject();
+  public lists: any[] = [];
+  public documentTableColumns: any[] = [
+    {
+      name: '★',
+      value: 'isFeatured',
+      width: 'col-1'
+    },
+    {
+      name: 'Name',
+      value: 'displayName',
+      width: 'col-3'
+    },
+    {
+      name: 'Date',
+      value: 'datePosted',
+      width: 'col-2'
+    },
+    {
+      name: 'Type',
+      value: 'type',
+      width: 'col-2'
+    },
+    {
+      name: 'Milestone',
+      value: 'milestone',
+      width: 'col-2'
+    },
+    {
+      name: 'Phase',
+      value: 'phase',
+      width: 'col-2'
+    }
+  ];
+
   private ngUnsubscribe: Subject<boolean> = new Subject<boolean>();
-  readonly defaultBounds = L.latLngBounds([48, -139], [60, -114]); // all of BC
-  private ngbModal: NgbModalRef = null;
 
   constructor(
+    private _changeDetectionRef: ChangeDetectorRef,
     private storageService: StorageService,
-    private elementRef: ElementRef,
-    public configService: ConfigService
+    public configService: ConfigService,
+    private tableTemplateUtils: TableTemplateUtils,
+    private searchService: SearchService,
+    private route: ActivatedRoute
   ) { }
 
   ngOnInit() {
     this.project = this.storageService.state.currentProject.data;
     this.commentPeriod = this.project.commentPeriodForBanner;
+    this.route.params
+    .subscribe(() => {
+      this.getFavoriteDocs();
+    });
+  }
+  // Featured Documents table addition
+
+  getFavoriteDocs() {
+    // Go to top of page after clicking to a different page.
+    this.loading = true;
+
+    this.tableParams = this.tableTemplateUtils.updateTableParams(this.tableParams, 1, this.tableParams.sortBy);
+
+    this.searchService.getFullList('List')
+        .takeUntil(this.ngUnsubscribe)
+        .subscribe((listRes: any) => {
+
+          this.lists = listRes[0].searchResults;
+
+          this.searchService.getSearchResults(
+            this.tableParams.keywords,
+            'Document',
+            [{ 'name': 'project', 'value': this.project._id }],
+            1,
+            this.tableParams.pageSize,
+            this.tableParams.sortBy,
+            { documentSource: 'PROJECT', isFeatured: 'true' },
+            true,
+            null,
+            {},
+            '')
+            .takeUntil(this.ngUnsubscribe)
+            .subscribe((res: any) => {
+              this.tableParams.totalListItems = res[0].data.searchResults.length;
+              this.featuredDocuments = res[0].data.searchResults;
+              this.tableTemplateUtils.updateUrl(this.tableParams.sortBy, this.tableParams.currentPage, this.tableParams.pageSize, null, this.tableParams.keywords);
+              this.setDocumentRowData();
+              this.loading = false;
+              this._changeDetectionRef.detectChanges();
+            });
+        });
   }
 
-  ngAfterViewInit() {
-    const self = this; // for closure function below
+  updateSelectedRow(count) {
+    this.selectedCount = count;
+  }
 
-    // custom control to reset map view
-    const resetViewControl = L.Control.extend({
-      options: {
-        position: 'topleft'
-      },
-      onAdd: function (map) {
-        const element = L.DomUtil.create('i', 'material-icons leaflet-bar leaflet-control leaflet-control-custom');
-
-        element.title = 'Reset view';
-        element.innerText = 'refresh'; // material icon name
-        element.style.width = '34px';
-        element.style.height = '34px';
-        element.style.lineHeight = '30px';
-        element.style.textAlign = 'center';
-        element.style.cursor = 'pointer';
-        element.style.backgroundColor = '#fff';
-        element.onmouseover = () => element.style.backgroundColor = '#f4f4f4';
-        element.onmouseout = () => element.style.backgroundColor = '#fff';
-
-        element.onclick = function () {
-          self.fitBounds(self.appFG.getBounds());
-        };
-
-        // prevent underlying map actions for these events
-        L.DomEvent.disableClickPropagation(element); // includes double-click
-        L.DomEvent.disableScrollPropagation(element);
-
-        return element;
-      },
-    });
-
-    // draw map
-    const Esri_OceanBasemap = L.tileLayer('https://server.arcgisonline.com/ArcGIS/rest/services/Ocean_Basemap/MapServer/tile/{z}/{y}/{x}', {
-      attribution: 'Tiles &copy; Esri &mdash; Sources: GEBCO, NOAA, CHS, OSU, UNH, CSUMB, National Geographic, DeLorme, NAVTEQ, and Esri',
-      maxZoom: 13,
-      noWrap: true
-    });
-    const Esri_NatGeoWorldMap = L.tileLayer('https://server.arcgisonline.com/ArcGIS/rest/services/NatGeo_World_Map/MapServer/tile/{z}/{y}/{x}', {
-      attribution: 'Tiles &copy; Esri &mdash; National Geographic, Esri, DeLorme, NAVTEQ, UNEP-WCMC, USGS, NASA, ESA, METI, NRCAN, GEBCO, NOAA, iPC',
-      maxZoom: 16,
-      noWrap: true
-    });
-    const OpenMapSurfer_Roads = L.tileLayer('https://korona.geog.uni-heidelberg.de/tiles/roads/x={x}&y={y}&z={z}', {
-      attribution: 'Imagery from <a href="http://giscience.uni-hd.de/">GIScience Research Group @ University of Heidelberg</a> &mdash; Map data &copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>',
-      maxZoom: 20,
-      noWrap: true
-    });
-    const World_Topo_Map = L.tileLayer('https://server.arcgisonline.com/ArcGIS/rest/services/World_Topo_Map/MapServer/tile/{z}/{y}/{x}', {
-      attribution: 'Tiles &copy; Esri &mdash; Esri, DeLorme, NAVTEQ, TomTom, Intermap, iPC, USGS, FAO, NPS, NRCAN, GeoBase, Kadaster NL, Ordnance Survey, Esri Japan, METI, Esri China (Hong Kong), and the GIS User Community',
-      maxZoom: 16,
-      noWrap: true
-    });
-    const World_Imagery = L.tileLayer('https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}', {
-      attribution: 'Tiles &copy; Esri &mdash; Source: Esri, i-cubed, USDA, USGS, AEX, GeoEye, Getmapping, Aerogrid, IGN, IGP, UPR-EGP, and the GIS User Community',
-      maxZoom: 17,
-      noWrap: true
-    });
-
-    this.map = L.map('map', {
-      zoomControl: false, // will be added manually below
-      maxBounds: L.latLngBounds(L.latLng(-90, -180), L.latLng(90, 180)), // restrict view to "the world"
-      zoomSnap: 0.1 // for greater granularity when fitting bounds
-    });
-
-    // NB: don't need to handle map change events
-    // since we always leave the subject app visible
-
-    // add reset view control
-    this.map.addControl(new resetViewControl());
-
-    // add zoom control
-    L.control.zoom({ position: 'topleft' }).addTo(this.map);
-
-    // add scale control
-    L.control.scale({ position: 'bottomright' }).addTo(this.map);
-
-    // add base maps layers control
-    const baseLayers = {
-      'Ocean Base': Esri_OceanBasemap,
-      'Nat Geo World Map': Esri_NatGeoWorldMap,
-      'Open Surfer Roads': OpenMapSurfer_Roads,
-      'World Topographic': World_Topo_Map,
-      'World Imagery': World_Imagery
-    };
-    L.control.layers(baseLayers).addTo(this.map);
-
-    // load base layer
-    for (const key of Object.keys(baseLayers)) {
-      if (key === this.configService.baseLayerName) {
-        this.map.addLayer(baseLayers[key]);
-        break;
-      }
+  setColumnSort(column) {
+    if (this.tableParams.sortBy.charAt(0) === '+') {
+      this.tableParams.sortBy = '-' + column;
+    } else {
+      this.tableParams.sortBy = '+' + column;
     }
+    this.getFavoriteDocs();
+  }
 
-    // save any future base layer changes
-    this.map.on('baselayerchange', function (e: L.LayersControlEvent) {
-      this.configService.baseLayerName = e.name;
-    }, this);
-
-    // Disable zoom on project details - iterferes with scrolling page.
-    this.map.scrollWheelZoom.disable();
-
-    // draw project marker
-    if (this.project) {
-      const markerIconYellow = L.icon({
-        iconUrl: 'assets/images/marker-icon-yellow.svg',
-        iconSize: [36, 36],
-        iconAnchor: [18, 36],
-        tooltipAnchor: [16, -28]
+  setDocumentRowData() {
+    let documentList = [];
+    if (this.featuredDocuments && this.featuredDocuments.length > 0) {
+      this.featuredDocuments.forEach(document => {
+        if (document) {
+          documentList.push(
+            {
+              documentFileName: document.documentFileName || document.displayName || document.internalOriginalName,
+              displayName: document.displayName,
+              datePosted: document.datePosted,
+              type: document.type,
+              milestone: document.milestone,
+              _id: document._id,
+              project: document.project,
+              isFeatured: document.isFeatured,
+              typeString: this.idToList(document.type),
+              milestoneString: this.idToList(document.milestone),
+              projectPhase: this.idToList(document.projectPhase)
+            }
+          );
+        }
       });
 
-      const title = `${this.project.name}\n`
-      + `${this.project.sector}\n`
-      + `${this.project.location}\n`;
-      const marker = L.marker(L.latLng(this.project.centroid[1], this.project.centroid[0]), { title: title })
-        .setIcon(markerIconYellow);
-      this.map.addLayer(marker);
+      this.documentTableData = new TableObject(
+        DocumentTableRowsComponent,
+        documentList,
+        this.tableParams
+      );
     }
-    this.map.addLayer(this.appFG);
-
-    this.fixMap();
   }
 
-  // to avoid timing conflict with animations (resulting in small map tile at top left of page),
-  // ensure map component is visible in the DOM then update it; otherwise wait a bit...
-  // ref: https://github.com/Leaflet/Leaflet/issues/4835
-  // ref: https://stackoverflow.com/questions/19669786/check-if-element-is-visible-in-dom
-  private fixMap() {
-    if (this.elementRef.nativeElement.offsetParent) {
-      this.fitBounds(this.appFG.getBounds());
+  idToList(id: string) {
+    if (!id) {
+      return '-';
+    }
+    // Grab the item from the constant lists, returning the name field of the object.
+    let item = this.lists.filter(listItem => listItem._id === id);
+    if (item.length !== 0) {
+      return item[0].name;
     } else {
-      setTimeout(this.fixMap.bind(this), 50);
+      return '-';
     }
-  }
-
-  public fitBounds(bounds: L.LatLngBounds = null) {
-    const fitBoundsOptions: L.FitBoundsOptions = {
-      // disable animation to prevent known bug where zoom is sometimes incorrect
-      // ref: https://github.com/Leaflet/Leaflet/issues/3249
-      animate: false,
-      // use bottom padding to keep shape in bounds
-      paddingBottomRight: [0, 35]
-    };
-
-    if (bounds && bounds.isValid()) {
-      this.map.fitBounds(bounds, fitBoundsOptions);
-    } else {
-      this.map.fitBounds(this.defaultBounds, fitBoundsOptions);
-    }
-  }
-
-  ngOnDestroy() {
-    if (this.ngbModal) { this.ngbModal.dismiss('component destroyed'); }
-    if (this.map) { this.map.remove(); }
-    this.ngUnsubscribe.next();
-    this.ngUnsubscribe.complete();
   }
 }
