@@ -4,13 +4,14 @@ import { Routes, RouterModule } from '@angular/router';
 import { ProjectComponent } from './project.component';
 import { ProjectResolver } from './project-resolver.service';
 import { ProjectRoutes } from './project-routes';
-import { CommentsComponent } from 'app/project/comments/comments.component';
-import { CommentsResolver } from 'app/project/comments/comments-resolver.service';
+import { CommentsComponent } from 'app/comments/comments.component';
+import { CommentsResolver } from 'app/comments/comments-resolver.service';
 import { DocumentsResolver } from './documents/documents-resolver.service';
 import { ProjectActivitiesResolver } from './project-activites/project-activities-resolver.service';
 import { DocumentTableResolver } from './documents/project-document-table-rows/project-document-table-rows-resolver.service';
 import { CertificatesResolver } from './certificates/certificates-resolver.service';
 import { AmendmentsResolverService } from './certificates/amendments-resolver.service';
+import { ApplicationResolver } from './certificates/application-resolver.service';
 import { PinsResolverService } from './pins/pins-resolver.service';
 
 const routes: Routes = [
@@ -54,7 +55,8 @@ const routes: Routes = [
     AmendmentsResolverService,
     ProjectActivitiesResolver,
     DocumentTableResolver,
-    PinsResolverService
+    PinsResolverService,
+    ApplicationResolver
   ]
 })
 
