@@ -13,12 +13,14 @@ import { SearchTerms } from 'app/models/search';
 import { TableObject } from 'app/shared/components/table-template/table-object';
 import { TableParamsObject } from 'app/shared/components/table-template/table-params-object';
 import { TableTemplateUtils } from 'app/shared/utils/table-template-utils';
+import { Constants } from 'app/shared/utils/constants';
 
 import { ProjectListTableRowsComponent } from './project-list-table-rows/project-list-table-rows.component';
 
 import { OrgService } from 'app/services/org.service';
 import { SearchService } from 'app/services/search.service';
 import { StorageService } from 'app/services/storage.service';
+import { ConfigService } from 'app/services/config.service';
 
 class ProjectFilterObject {
   constructor(
@@ -30,8 +32,9 @@ class ProjectFilterObject {
     public proponent: Array<Org> = [],
     public region: Array<string> = [],
     public CEAAInvolvement: Array<string> = [],
+    public projectPhase: Array<string> = [],
     public vc: Array<object> = []
-  ) {}
+  ) { }
 }
 
 @Component({
@@ -40,10 +43,15 @@ class ProjectFilterObject {
   styleUrls: ['./project-list.component.scss']
 })
 export class ProjectListComponent implements OnInit, OnDestroy {
+  public readonly constants = Constants;
   public projects: Array<Project> = [];
   public proponents: Array<Org> = [];
   public regions: Array<object> = [];
   public ceaaInvolvements: Array<object> = [];
+  public eacDecisions: Array<any> = [];
+  public commentPeriods: Array<object> = [];
+  public projectTypes: Array<object> = [];
+  public projectPhases: Array<object> = [];
 
   public loading = true;
 
@@ -107,81 +115,8 @@ export class ProjectListComponent implements OnInit, OnDestroy {
 
   private ngUnsubscribe: Subject<boolean> = new Subject<boolean>();
 
-  private TYPE_MAP: object = {
-    energyElectricity: 'Energy-Electricity',
-    energyPetroleum: 'Energy-Petroleum & Natural Gas',
-    foodProcessing: 'Food Processing',
-    industrial: 'Industrial',
-    mines: 'Mines',
-    other: 'Other',
-    tourist: 'Tourist Destination Resorts',
-    transportation: 'Transportation',
-    wasteDisposal: 'Waste Disposal',
-    waterManagement: 'Water Management'
-  };
-
-  private EAC_DECISIONS_MAP: object = {
-    inProgress: 'In Progress',
-    certificateIssued: 'Certificate Issued',
-    certificateRefused: 'Certificate Refused',
-    furtherAssessmentRequired: 'Further Assessment Required',
-    certificateNotRequired: 'Certificate Not Required',
-    certificateExpired: 'Certificate Expired',
-    withdrawn: 'Withdrawn',
-    terminated: 'Terminated',
-    preEA: 'Pre-EA Act Approval',
-    notReviewable: 'Not Designated Reviewable'
-  };
-
-  private PCP_MAP: object = {
-    pending: 'pending',
-    open: 'open',
-    closed: 'closed'
-  };
-
-  private REGIONS_COLLECTION: Array<object> = [
-    { code: 'Cariboo', name: 'Cariboo' },
-    { code: 'Kootenay', name: 'Kootenay' },
-    { code: 'Lower Mainland', name: 'Lower Mainland' },
-    { code: 'Okanagan', name: 'Okanagan' },
-    { code: 'Omineca', name: 'Omineca' },
-    { code: 'Peace', name: 'Peace' },
-    { code: 'Skeena', name: 'Skeena' },
-    { code: 'Thompson-Nicola', name: 'Thompson-Nicola' },
-    { code: 'Vancouver Island', name: 'Vancouver Island' }
-  ];
-
-  private CEAA_INVOLVEMENTS_COLLECTION: Array<object> = [
-    { code: 'None', name: 'None' },
-    { code: 'Panel', name: 'Panel' },
-    { code: 'Panel (CEAA 2012)', name: 'Panel (CEAA 2012)' },
-    { code: 'Coordinated', name: 'Coordinated' },
-    { code: 'Screening', name: 'Screening' },
-    { code: 'Screening - Confirmed', name: 'Screening - Confirmed' },
-    { code: 'Substituted', name: 'Substituted' },
-    {
-      code: 'Substituted (Provincial Lead)',
-      name: 'Substituted (Provincial Lead)'
-    },
-    { code: 'Comprehensive Study', name: 'Comprehensive Study' },
-    {
-      code: 'Comprehensive Study - Unconfirmed',
-      name: 'Comprehensive Study - Unconfirmed'
-    },
-    {
-      code: 'Comprehensive Study - Confirmed',
-      name: 'Comprehensive Study - Confirmed'
-    },
-    {
-      code: 'Comprehensive Study (Pre CEAA 2012)',
-      name: 'Comprehensive Study (Pre CEAA 2012)'
-    },
-    { code: 'Comp Study', name: 'Comp Study' },
-    { code: 'Comp Study - Unconfirmed', name: 'Comp Study - Unconfirmed' },
-    { code: 'To be determined', name: 'To be determined' },
-    { code: 'Equivalent - NEB', name: 'Equivalent - NEB' },
-    { code: 'Yes', name: 'Yes' }
-  ];
+  private previousFilters;
+  private previousKeyword;
 
   constructor(
     private router: Router,
@@ -190,82 +125,135 @@ export class ProjectListComponent implements OnInit, OnDestroy {
     private storageService: StorageService,
     private searchService: SearchService,
     private orgService: OrgService,
+    private config: ConfigService,
     private _changeDetectionRef: ChangeDetectorRef
-  ) {}
+  ) { }
 
   ngOnInit() {
-    this.route.params.takeUntil(this.ngUnsubscribe).subscribe(params => {
-      let newParams = params;
-
-      if (
-        Object.keys(newParams).length === 0 &&
-        newParams.constructor === Object
-      ) {
-        newParams = {
-          sortBy: '+name'
-        };
-      }
-
-      // Fetch proponents
-      this.orgService
-        .getByCompanyType('Proponent/Certificate Holder')
-        .takeUntil(this.ngUnsubscribe)
-        .subscribe((proponents: any) => {
-          if (proponents) {
-            this.proponents = proponents;
-
-            this.regions = this.REGIONS_COLLECTION;
-            this.ceaaInvolvements = this.CEAA_INVOLVEMENTS_COLLECTION;
-
-            this.setFiltersFromParams(params);
-
-            this.updateCounts();
-
-            this.tableParams = this.tableTemplateUtils.getParamsFromUrl(
-              newParams,
-              this.filterForURL
-            );
-
-            this.searchService
-              .getSearchResults(
-                this.tableParams.keywords,
-                'Project',
-                [],
-                this.tableParams.currentPage,
-                this.tableParams.pageSize,
-                this.tableParams.sortBy,
-                {},
-                true,
-                null,
-                this.filterForAPI,
-                ''
-              )
-              .takeUntil(this.ngUnsubscribe)
-              .subscribe((res: any) => {
-                if (res[0].data) {
-                  if (res[0].data.searchResults.length > 0) {
-                    this.tableParams.totalListItems =
-                      res[0].data.meta[0].searchResultsTotal;
-                    this.projects = res[0].data.searchResults;
-                  } else {
-                    this.tableParams.totalListItems = 0;
-                    this.projects = [];
-                  }
-                  this.setRowData();
-                } else {
-                  alert('Uh-oh, couldn\'t load search results');
-                  // results not found --> navigate back to search
-                  this.router.navigate(['/']);
-                }
-                this.loading = false;
-                this._changeDetectionRef.detectChanges();
-              });
-          } else {
-            alert('Uh-oh, couldn\'t load proponents');
-            this.router.navigate(['/']);
+    // Fetch proponents
+    this.orgService
+      .getByCompanyType('Proponent/Certificate Holder')
+      .switchMap((res: any) => {
+        if (res) {
+          this.proponents = res || [];
+          return this.config.lists;
+        } else {
+          alert('Uh-oh, couldn\'t load proponents');
+          this.router.navigate(['/']);
+        }
+      })
+      .switchMap((list: any) => {
+        list.map(item => {
+          switch (item.type) {
+            case 'eaDecisions':
+              this.eacDecisions.push({ ...item });
+              break;
+            case 'ceaaInvolvements':
+              this.ceaaInvolvements.push({ ...item });
+              break;
+            case 'projectPhase':
+              this.projectPhases.push({ ...item});
+              break;
           }
         });
-    });
+
+        this.regions = Constants.REGIONS_COLLECTION;
+        this.commentPeriods = Constants.PCP_COLLECTION;
+        this.projectTypes = Constants.PROJECT_TYPE_COLLECTION;
+
+        return this.route.params;
+      })
+      .switchMap((res: any) => {
+        let params = { ...res };
+
+        this.setFiltersFromParams(params);
+
+        // default sort for project list is alphabetical
+        params.sortBy = '+name'
+
+        this.updateCounts();
+
+        this.tableParams = this.tableTemplateUtils.getParamsFromUrl(
+          params,
+          this.filterForURL
+        );
+
+        if (this.tableParams.sortBy === '') {
+          this.tableParams.sortBy = '+name';
+        }
+
+        // check if the filters are in session state, for handling
+        // retaining the filters when a user clicks back from a project
+        // into the project list
+        if (this.storageService && this.storageService.state.projList) {
+          this.filterForAPI = this.storageService.state.projList.filterForAPI;
+          this.filterForUI = this.storageService.state.projList.filterForUI;
+          this.tableParams = this.storageService.state.projList.tableParams;
+          this.setParamsFromFilters(params);
+        }
+
+        if (this.filterForAPI.hasOwnProperty('projectPhase')) {
+          this.filterForAPI['currentPhaseName'] = this.filterForAPI['projectPhase'];
+          delete this.filterForAPI['projectPhase'];
+        }
+
+        return this.searchService
+          .getSearchResults(
+            this.tableParams.keywords,
+            'Project',
+            [],
+            this.tableParams.currentPage,
+            this.tableParams.pageSize,
+            this.tableParams.sortBy,
+            {},
+            true,
+            null,
+            this.filterForAPI,
+            ''
+          );
+      })
+      .takeUntil(this.ngUnsubscribe)
+      .subscribe((res: any) => {
+
+        if (this.filterForAPI.hasOwnProperty('currentPhaseName')) {
+          this.filterForAPI['projectPhase'] = this.filterForAPI['currentPhaseName'];
+          delete this.filterForAPI['currentPhaseName'];
+        }
+
+        if (res[0].data) {
+          if (res[0].data.searchResults.length > 0) {
+            this.tableParams.totalListItems =
+              res[0].data.meta[0].searchResultsTotal;
+            this.projects = res[0].data.searchResults;
+          } else {
+            this.tableParams.totalListItems = 0;
+            this.projects = [];
+          }
+          this.setRowData();
+
+          // store the state of the filterForAPI set into the session
+          // so a user can navigate back to this page without losing
+          // their filters
+          if (this.storageService) {
+            this.storageService.state.projList = {};
+            this.storageService.state.projList.filterForAPI = this.filterForAPI;
+            this.storageService.state.projList.filterForUI = this.filterForUI;
+            this.storageService.state.projList.tableParams = this.tableParams;
+          }
+
+          // We need to clone filters, not reference
+          this.previousFilters = { ...this.filterForAPI };
+          this.previousKeyword = this.tableParams.keywords;
+
+        } else {
+          alert('Uh-oh, couldn\'t load search results');
+          // results not found --> navigate back to search
+          this.router.navigate(['/']);
+        }
+        this.loading = false;
+        this._changeDetectionRef.detectChanges();
+      });
+
   }
 
   addProject() {
@@ -297,7 +285,6 @@ export class ProjectListComponent implements OnInit, OnDestroy {
   }
 
   paramsToCollectionFilters(params, name, collection, identifyBy) {
-    this.filterForUI[name] = [];
     delete this.filterForURL[name];
     delete this.filterForAPI[name];
 
@@ -306,9 +293,8 @@ export class ProjectListComponent implements OnInit, OnDestroy {
       // look up each value in collection
       const values = params[name].split(',');
       values.forEach(value => {
-        const record = _.find(collection, [identifyBy, value]);
+        const record = _.find(collection, [ identifyBy, value ]);
         if (record) {
-          this.filterForUI[name].push(record);
           confirmedValues.push(value);
         }
       });
@@ -338,19 +324,14 @@ export class ProjectListComponent implements OnInit, OnDestroy {
   }
 
   setFiltersFromParams(params) {
-    this.paramsToCheckboxFilters(params, 'type', this.TYPE_MAP);
-    this.paramsToCheckboxFilters(params, 'eacDecision', this.EAC_DECISIONS_MAP);
-    this.paramsToCheckboxFilters(params, 'pcp', this.PCP_MAP);
-
+    this.paramsToCollectionFilters(params, 'type', this.projectTypes, 'name');
+    this.paramsToCollectionFilters(params, 'pcp', this.commentPeriods, 'code');
     this.paramsToCollectionFilters(params, 'region', this.regions, 'code');
-    this.paramsToCollectionFilters(
-      params,
-      'CEAAInvolvement',
-      this.ceaaInvolvements,
-      'code'
-    );
+    this.paramsToCollectionFilters(params, 'CEAAInvolvement', this.ceaaInvolvements, '_id');
     this.paramsToCollectionFilters(params, 'proponent', this.proponents, '_id');
     this.paramsToCollectionFilters(params, 'vc', null, '_id');
+    this.paramsToCollectionFilters(params, 'eacDecision', this.eacDecisions, '_id');
+    this.paramsToCollectionFilters(params, 'projectPhase', this.projectPhases, '_id');
 
     this.paramsToDateFilters(params, 'decisionDateStart');
     this.paramsToDateFilters(params, 'decisionDateEnd');
@@ -393,14 +374,14 @@ export class ProjectListComponent implements OnInit, OnDestroy {
   }
 
   setParamsFromFilters(params) {
-    this.checkboxFilterToParams(params, 'type');
-    this.checkboxFilterToParams(params, 'eacDecision');
-    this.checkboxFilterToParams(params, 'pcp');
-
+    this.collectionFilterToParams(params, 'type', 'name');
+    this.collectionFilterToParams(params, 'pcp', 'code');
+    this.collectionFilterToParams(params, 'eacDecision', '_id');
     this.collectionFilterToParams(params, 'region', 'code');
-    this.collectionFilterToParams(params, 'CEAAInvolvement', 'code');
+    this.collectionFilterToParams(params, 'CEAAInvolvement', '_id');
     this.collectionFilterToParams(params, 'proponent', '_id');
     this.collectionFilterToParams(params, 'vc', '_id');
+    this.collectionFilterToParams(params, 'projectPhase', '_id');
 
     this.dateFilterToParams(params, 'decisionDateStart');
     this.dateFilterToParams(params, 'decisionDateEnd');
@@ -428,6 +409,8 @@ export class ProjectListComponent implements OnInit, OnDestroy {
   }
 
   clearAll() {
+    this.tableParams.keywords = '';
+    delete this.filterForURL['keywords'];
     Object.keys(this.filterForUI).forEach(key => {
       if (this.filterForUI[key]) {
         if (Array.isArray(this.filterForUI[key])) {
@@ -487,12 +470,12 @@ export class ProjectListComponent implements OnInit, OnDestroy {
           eacDecision: project.eacDecision
         });
       });
-      this.projectTableData = new TableObject(
-        ProjectListTableRowsComponent,
-        projectList,
-        this.tableParams
-      );
     }
+    this.projectTableData = new TableObject(
+      ProjectListTableRowsComponent,
+      projectList,
+      this.tableParams
+    );
   }
 
   setColumnSort(column) {
@@ -515,6 +498,36 @@ export class ProjectListComponent implements OnInit, OnDestroy {
       this.tableParams.sortBy
     );
 
+    // the need to do this here is mildly disturbing
+    // TODO: replace the param-filter calls
+    let params = this.terms.getParams();
+    params['ms'] = new Date().getMilliseconds();
+    params['dataset'] = this.terms.dataset;
+    params['currentPage'] = this.tableParams.currentPage;
+    params['sortBy'] = this.tableParams.sortBy;
+    params['keywords'] = this.tableParams.keywords;
+    params['pageSize'] = this.tableParams.pageSize
+    this.setParamsFromFilters(params);
+    this.setFiltersFromParams(params);
+
+    // store the table params in the event of a page navigation
+    this.storageService.state.projList.tableParams = this.tableParams;
+
+    if (this.filterForAPI.hasOwnProperty('projectPhase')) {
+      this.filterForAPI['currentPhaseName'] = this.filterForAPI['projectPhase'];
+      delete this.filterForAPI['projectPhase'];
+    }
+
+    this.storageService.state.projList.filterForAPI = this.filterForAPI;
+    this.storageService.state.projList.filterForUI = this.filterForUI;
+    this.storageService.state.projList.tableParams = this.tableParams;
+
+    if (this.tableParams.keywords !== this.previousKeyword || JSON.stringify(this.filterForAPI) !== JSON.stringify(this.previousFilters)) {
+      this.projectTableData.paginationData.currentPage = 1;
+      this.tableParams.currentPage = 1;
+      pageNumber = 1;
+    }
+
     this.searchService
       .getSearchResults(
         this.tableParams.keywords,
@@ -532,8 +545,9 @@ export class ProjectListComponent implements OnInit, OnDestroy {
       .takeUntil(this.ngUnsubscribe)
       .subscribe((res: any) => {
         if (res && res[0].data) {
-          this.tableParams.totalListItems =
-            res[0].data.meta[0].searchResultsTotal;
+          this.tableParams.totalListItems = res[0].data.searchResults.length > 0
+                                          ? res[0].data.meta[0].searchResultsTotal
+                                          : 0;
           this.projects = res[0].data.searchResults;
           this.tableTemplateUtils.updateUrl(
             this.tableParams.sortBy,
@@ -545,22 +559,24 @@ export class ProjectListComponent implements OnInit, OnDestroy {
           this.setRowData();
           this.loading = false;
           this._changeDetectionRef.detectChanges();
+
+          this.previousFilters = { ...this.filterForAPI };
+          this.previousKeyword = this.tableParams.keywords;
+
         } else {
-          alert('Uh-oh, couldn\'t load topics');
+          alert('Uh-oh, couldn\'t load projects');
           // project not found --> navigate back to search
           this.router.navigate(['/']);
         }
       });
+
+      if (this.filterForAPI.hasOwnProperty('currentPhaseName')) {
+        this.filterForAPI['projectPhase'] = this.filterForAPI['currentPhaseName'];
+        delete this.filterForAPI['currentPhaseName'];
+      }
   }
 
-  public onSubmit(currentPage = 1) {
-    // dismiss any open snackbar
-    // if (this.snackBarRef) { this.snackBarRef.dismiss(); }
-
-    // NOTE: Angular Router doesn't reload page on same URL
-    // REF: https://stackoverflow.com/questions/40983055/how-to-reload-the-current-route-with-the-angular-2-router
-    // WORKAROUND: add timestamp to force URL to be different than last time
-
+  public onSubmit() {
     let params = this.terms.getParams();
     params['ms'] = new Date().getMilliseconds();
     params['dataset'] = this.terms.dataset;
@@ -573,6 +589,19 @@ export class ProjectListComponent implements OnInit, OnDestroy {
 
     console.log('onSubmit params', params);
     this.router.navigate(['projects-list', params]);
+  }
+
+  // Compares selected options when a dropdown is grouped by legislation.
+  compareDropdownOptions(optionA: any, optionB: any) {
+    if ((optionA.name === optionB.name) && (optionA.legislation === optionB.legislation)) {
+      return true;
+    }
+
+    return false;
+  }
+
+  clearSelectedItem(filter: string, item: any) {
+    this.filterForUI[filter] = this.filterForUI[filter].filter(option => option._id !== item._id);
   }
 
   ngOnDestroy() {
