@@ -2,8 +2,9 @@ import { async, TestBed } from '@angular/core/testing';
 import { Decision } from 'app/models/decision';
 import 'rxjs/add/observable/of';
 import { Observable } from 'rxjs/Observable';
+import 'rxjs/add/observable/throw';
 import { ApiService } from './api';
-import { DecisionService } from './Decision.service';
+import { DecisionService } from './decision.service';
 import { DocumentService } from './document.service';
 import { Document } from 'app/models/document';
 
