@@ -4,6 +4,7 @@ export class TableObject {
   constructor(
     public component: Type<any>,
     public data: any,
-    public paginationData: any = null
+    public paginationData: any = null,
+    public extraData: any = null
   ) { }
 }
