@@ -11,14 +11,16 @@ import { ObjectFilterPipe } from 'app/shared/pipes/object-filter.pipe';
 import { VarDirective } from 'app/shared/utils/ng-var.directive';
 import { DragMoveDirective } from 'app/shared/utils/drag-move.directive';
 
-import { NgxPaginationModule } from 'ngx-pagination';
 import { TableTemplateComponent } from 'app/shared/components/table-template/table-template.component';
 import { ListConverterPipe } from './pipes/list-converter.pipe';
+import { SafeHtmlPipe } from './pipes/safe-html-converter.pipe';
 import { OrgNamePipe } from './pipes/org-name.pipe';
 import { TableTemplateUtils } from './utils/table-template-utils';
 import { TableDirective } from './components/table-template/table.directive';
 import { PublishedPipe } from 'app/shared/pipes/published.pipe';
 import { Utils } from './utils/utils';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgZorroAntdModule, NZ_I18N, en_US } from 'ng-zorro-antd';
 
 @NgModule({
   imports: [
@@ -26,7 +28,8 @@ import { Utils } from './utils/utils';
     MatProgressBarModule,
     MatSnackBarModule,
     NgxTextOverflowClampModule,
-    NgxPaginationModule
+    BrowserAnimationsModule,
+    NgZorroAntdModule
   ],
   declarations: [
     OrderByPipe,
@@ -38,6 +41,7 @@ import { Utils } from './utils/utils';
     TableTemplateComponent,
     TableDirective,
     ListConverterPipe,
+    SafeHtmlPipe,
     OrgNamePipe
   ],
   exports: [
@@ -52,13 +56,14 @@ import { Utils } from './utils/utils';
     VarDirective,
     DragMoveDirective,
     TableTemplateComponent,
-    NgxPaginationModule,
     ListConverterPipe,
+    SafeHtmlPipe,
     OrgNamePipe
   ],
   providers: [
     TableTemplateUtils,
-    Utils
+    Utils,
+    { provide: NZ_I18N, useValue: en_US }
   ]
 })
 
