@@ -19,21 +19,21 @@
  */
 
 /** IE9, IE10 and IE11 requires all of the following polyfills. **/
-import 'core-js/es6/symbol';
-import 'core-js/es6/object';
-import 'core-js/es6/function';
-import 'core-js/es6/parse-int';
-import 'core-js/es6/parse-float';
-import 'core-js/es6/number';
-import 'core-js/es6/math';
-import 'core-js/es6/string';
-import 'core-js/es6/date';
-import 'core-js/es6/array';
-import 'core-js/es6/regexp';
-import 'core-js/es6/map';
-import 'core-js/es6/set';
-import 'core-js/es6/reflect';
-import 'core-js/client/shim';
+import 'core-js/es/symbol';
+import 'core-js/es/object';
+import 'core-js/es/function';
+import 'core-js/es/parse-int';
+import 'core-js/es/parse-float';
+import 'core-js/es/number';
+import 'core-js/es/math';
+import 'core-js/es/string';
+import 'core-js/es/date';
+import 'core-js/es/array';
+import 'core-js/es/regexp';
+import 'core-js/es/map';
+import 'core-js/es/set';
+import 'core-js/es/reflect';
+import 'core-js/features/reflect';
 
 /** IE10 and IE11 requires the following for NgClass support on SVG elements */
 import 'classlist.js';  // Run `npm install --save classlist.js`.
@@ -43,8 +43,8 @@ import 'web-animations-js';  // Run `npm install --save web-animations-js`.
 
 
 /** Evergreen browsers require these. **/
-import 'core-js/es7/reflect';
-import 'core-js/es7/array';
+import 'core-js/es/reflect';
+import 'core-js/es/array';
 
 
 
@@ -64,14 +64,8 @@ import 'zone.js/dist/zone';  // Included with Angular CLI.
  * APPLICATION IMPORTS
  */
 
-/**
- * Date, currency, decimal and percent pipes.
- * Needed for: All but Chrome, Firefox, Edge, IE11 and Safari 10
- */
-import 'intl';  // Run `npm install --save intl`.
-import 'intl/locale-data/complete.js';
-import 'intl/locale-data/jsonp/en.js';
-
+// global is not defined fix (https://github.com/Leonidas-from-XIV/node-xml2js/issues/538)
+(window as any)['global'] = window;
 
 // Specific polyfill needed for IE11 using current packages as of Sep 24, 2018:
 // https://github.com/uxitten/polyfill/blob/master/string.polyfill.js
